Move selectedRefundForDetail state to parent as binding
